Umount disk but don't reboot machine
It is required to either patch the VM with a lower bootOrder for the
disks to use with OS OR remove the disk with virtualMedia, choosing the
later since is already implemented, but not rebooting the VM. Changes
will be reflected on the next boot.

https://kubevirt.io/api-reference/v1.1.1/definitions.html#_v1_disk
BootOrder is an integer value > 0, used to determine ordering of boot
devices. Lower values take precedence. Each disk or interface that has
a boot order must have a unique value. Disks without a boot order are
not tried if a disk with a boot order exists.